+++ site/controller.php
@@ -85,7 +85,7 @@ class GetbibleController extends BaseController
 		if (StringHelper::check($view))
 		{
 			$views = array(
-
+				''
 				);
 			// check if this is a edit view
 			if (in_array($view,$views))